Add command line options to include contextual information
Currently the code in Testomatio only shows the scenario's source code.
However, the descriptions of parent features and rules as well as
the steps of background sections may also be relevant to the tester.
Thus, this PR adds command line options to include the source code
of these elemnts in the exported scenatios.